List mnesia as dependency of test application
An OTP release using mnesia sumo_db adapter must list mnesia as a
dependency - either in one of the applications of the OTP release or
in the rebar.config - because neither
[sumo_rest](https://github.com/inaka/sumo_rest/blob/0.3.3/src/sumo_rest.app.src)
(of course) nor
[sumo_db](https://github.com/inaka/sumo_db/blob/0.7.1/src/sumo_db.app.src)
depend on mnesia.

It makes sense that sumo_db does not depend on mnesia, as its users do
not necessarily use the included mnesia adapter.

As the test application does not have its own rebar.config, list
mnesia as dependency of the application sr_test so that novice sumo_db
user is reminded to list it.